add validate.js
