+++ task-1/styles/noteModal.css
@@ -1,6 +1,3 @@
-/* styles.css або modal.css */
-
-/* Стилі для модального вікна */
 .modal {
     display: none;
     position: fixed;
@@ -8,7 +5,7 @@
     left: 0;
     width: 100%;
     height: 100%;
-    background-color: rgba(0, 0, 0, 0.5); /* Прозорість тла */
+    background-color: var(--bg-modal);
 }
 
 .modal-content {
@@ -101,5 +98,5 @@
 }
 
 .highlight {
-    border: 1px solid red;
+    border: 1px solid var(--error);
 }
